revert: "fix: Add new route to create team routing-form response (#22347)" (#22399)
This reverts commit 89e7fca1839f4328937d895516b6cd14ce847a21.
