--- apps/api/v2/src/modules/auth/guards/or-guard/index.ts
@@ -1,1 +1,0 @@
-export { Or } from "./or.guard";
