+++ apps/api/v2/src/modules/organizations/routing-forms/organizations-routing-forms.module.ts
@@ -1,5 +1,3 @@
-import { EventTypesRepository_2024_06_14 } from "@/ee/event-types/event-types_2024_06_14/event-types.repository";
-import { IsUserRoutingForm } from "@/modules/auth/guards/organizations/is-user-routing-form.guard";
 import { MembershipsRepository } from "@/modules/memberships/memberships.repository";
 import { OrganizationsRepository } from "@/modules/organizations/index/organizations.repository";
 import { OrganizationsTeamsRoutingFormsResponsesOutputService } from "@/modules/organizations/teams/routing-forms/services/organizations-teams-routing-forms-responses-output.service";
@@ -16,21 +14,17 @@ import { OrganizationsRoutingFormsController } from "./controllers/organizations
 import { OrganizationsRoutingFormsRepository } from "./organizations-routing-forms.repository";
 import { OrganizationsRoutingFormsResponsesService } from "./services/organizations-routing-forms-responses.service";
 import { OrganizationsRoutingFormsService } from "./services/organizations-routing-forms.service";
-import { SharedRoutingFormResponseService } from "./services/shared-routing-form-response.service";
 
 @Module({
   imports: [PrismaModule, StripeModule, RedisModule, RoutingFormsModule, SlotsModule_2024_09_04],
   providers: [
-    IsUserRoutingForm,
     MembershipsRepository,
     OrganizationsRepository,
     OrganizationsRoutingFormsRepository,
     OrganizationsRoutingFormsService,
     OrganizationsRoutingFormsResponsesService,
-    SharedRoutingFormResponseService,
     OrganizationsTeamsRoutingFormsResponsesOutputService,
     TeamsEventTypesRepository,
-    EventTypesRepository_2024_06_14,
   ],
   controllers: [OrganizationsRoutingFormsController, OrganizationsRoutingFormsResponsesController],
 })
