+++ apps/api/v2/src/modules/organizations/teams/routing-forms/organizations-teams-routing-forms.module.ts
@@ -1,9 +1,7 @@
-import { EventTypesRepository_2024_06_14 } from "@/ee/event-types/event-types_2024_06_14/event-types.repository";
 import { MembershipsRepository } from "@/modules/memberships/memberships.repository";
 import { OrganizationsRepository } from "@/modules/organizations/index/organizations.repository";
 import { OrganizationsRoutingFormsRepository } from "@/modules/organizations/routing-forms/organizations-routing-forms.repository";
 import { OrganizationsRoutingFormsResponsesService } from "@/modules/organizations/routing-forms/services/organizations-routing-forms-responses.service";
-import { SharedRoutingFormResponseService } from "@/modules/organizations/routing-forms/services/shared-routing-form-response.service";
 import { OrganizationsTeamsRepository } from "@/modules/organizations/teams/index/organizations-teams.repository";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
 import { RedisModule } from "@/modules/redis/redis.module";
@@ -33,7 +31,6 @@ import { OrganizationsTeamsRoutingFormsService } from "./services/organizations-
   providers: [
     OrganizationsTeamsRoutingFormsService,
     OrganizationsTeamsRoutingFormsResponsesService,
-    SharedRoutingFormResponseService,
     OrganizationsTeamsRoutingFormsResponsesOutputService,
     OrganizationsTeamsRoutingFormsResponsesRepository,
     OrganizationsTeamsRoutingFormsRepository,
@@ -42,7 +39,6 @@ import { OrganizationsTeamsRoutingFormsService } from "./services/organizations-
     MembershipsRepository,
     OrganizationsRoutingFormsResponsesService,
     OrganizationsRoutingFormsRepository,
-    EventTypesRepository_2024_06_14,
   ],
   controllers: [OrganizationsTeamsRoutingFormsResponsesController, OrganizationsTeamsRoutingFormsController],
 })
